+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # Changelog
 
 
+## 1.1.2
+ * 完善了input widget验证功能.
+
 ## 1.1.1
- * 进一步完善了input widget.
+ * 完善了input widget.
 
 ## 1.1.0
  * 更新了button的风格样式并生成了相应的widget类.